--- a/spec/support/jobs/concerns/etl/identity_providers.rb
+++ b/spec/support/jobs/concerns/etl/identity_providers.rb
@@ -2,7 +2,6 @@
 
 RSpec.shared_examples 'ETL::IdentityProviders' do
   include_examples 'ETL::Common'
-  # rubocop:disable Metrics/MethodLength,Metrics/AbcSize
   def create_json(idp)
     contact_people =
       contact_instances.map { |cp| contact_person_json(cp) } +
@@ -51,7 +50,7 @@ RSpec.shared_examples 'ETL::IdentityProviders' do
       }
     }
   end
-  # rubocop:enable Metrics/MethodLength,Metrics/AbcSize
+  # rubocop:enable
 
   def attribute_json(a)
     {
@@ -207,16 +206,8 @@ RSpec.shared_examples 'ETL::IdentityProviders' do
       context 'scopes' do
         it 'sets a scope' do
           expect(subject.scopes.size).to eq(1)
-        end
-
-        it 'sets expected scope' do
           expect(subject.scopes.first.value).to eq(scope)
-        end
-
-        context 'normal scope' do
-          it 'sets regex to false' do
-            expect(subject.scopes.first.regexp).not_to be
-          end
+          expect(subject.scopes.first.regexp).not_to be
         end
 
         context 'regex scope' do
@@ -331,41 +322,22 @@ RSpec.shared_examples 'ETL::IdentityProviders' do
     include_examples 'updating an SSODescriptor'
     include_examples 'updating MDUI content'
 
-    it 'uses the existing instance' do
-      expect { run }.not_to(change { IDPSSODescriptor.count })
-    end
-
-    it 'sets a scope' do
+    it 'Works as expected' do
+      expect { run }.to(not_change { IDPSSODescriptor.count }.and(
+        not_change { Tag.count }
+      ).and(
+        change { subject.reload.single_sign_on_services }
+      ).and(
+        change { subject.reload.name_id_mapping_services }
+      ).and(
+        change { subject.reload.assertion_id_request_services }
+      ).and(
+        change { subject.reload.attribute_profiles }
+      ).and(
+        change { subject.reload.attributes }
+      ))
       expect(subject.scopes.size).to eq(1)
-    end
-
-    it 'sets expected scope' do
       expect(subject.scopes.first.value).to eq(scope)
-    end
-
-    it 'does not create more tags' do
-      expect { run }.not_to(change { Tag.count })
-    end
-
-    it 'updates single sign on services' do
-      expect { run }.to(change { subject.reload.single_sign_on_services })
-    end
-
-    it 'updates name id mapping services' do
-      expect { run }.to(change { subject.reload.name_id_mapping_services })
-    end
-
-    it 'updates assertion id request services' do
-      expect { run }
-        .to(change { subject.reload.assertion_id_request_services })
-    end
-
-    it 'updates attribute profiles' do
-      expect { run }.to(change { subject.reload.attribute_profiles })
-    end
-
-    it 'updates attributes' do
-      expect { run }.to(change { subject.reload.attributes })
     end
   end