… .github/workflows/deploy.yml
… spec/support/jobs/concerns/etl/saml_uri.rb
… spec/support/jobs/concerns/etl/service_providers.rb
… spec/support/metadata/attribute.rb
… spec/support/metadata/attribute_authority_descriptor.rb
… spec/support/metadata/attribute_consuming_service.rb
… spec/support/metadata/contact_person.rb
… spec/support/metadata/endpoint.rb
… spec/support/metadata/entities_descriptor.rb
… spec/support/metadata/entities_descriptor_xml.rb
… spec/support/metadata/entity_descriptor.rb
… spec/support/metadata/idp_sso_descriptor.rb
… spec/support/metadata/key_descriptor.rb
… spec/support/metadata/key_info.rb
… spec/support/metadata/mdattr_entity_attribute.rb
… spec/support/metadata/mdrpi_publisher_info.rb
… spec/support/metadata/mdrpi_registration_info.rb
… spec/support/metadata/mdui_disco_hints.rb
… spec/support/metadata/mdui_ui_info.rb
… spec/support/metadata/organization.rb
… spec/support/metadata/raw_entity_descriptor.rb
… spec/support/metadata/requested_attribute.rb
… spec/support/metadata/role_descriptor.rb
… spec/support/metadata/saml_namespaces.rb
… spec/support/metadata/shibmd_key_authority.rb
… spec/support/metadata/shibmd_scope.rb
… spec/support/metadata/signature.rb
… spec/support/metadata/sirtfi_contact_person.rb
… spec/support/metadata/sp_sso_descriptor.rb
… spec/support/metadata/sso_role_descriptor.rb