(ignore the failing tests for now)
This was from a run reducing just the ETL tests, and there is still improvements that could be done by chaining .and most are just merged it blocks where context is no different between tests.

The one downside to this change is the context/it blocks that are removed that may provide more information to a dev.

While this time taken might not feel like much, it would proably get better by performing such adjustments across the whole test suite, also locally for some reason the entire suite can take ~25 minutes on my machine so hoping these changes improve this also

Before:
Finished in 4 minutes 11.3 seconds (files took 4.85 seconds to load)
2098 examples, 0 failures
After:
Finished in 3 minutes 46.7 seconds (files took 6.48 seconds to load)
1654 examples, 0 failures